+++ app/src/main/java/com/example/tryuserapp/tools/FirebaseHelper.kt
@@ -33,7 +33,7 @@ class FirebaseHelper {
                 idHotel = queryDocumentSnapshot.getString("idHotel") ?: "",
                 namaKatalis = queryDocumentSnapshot.getString("namaKatalis") ?: "",
                 imageLink = queryDocumentSnapshot.getString("imageLink")?: "",
-                stokKatalis = queryDocumentSnapshot.getLong("stokKatalis")?.toFloat() ?: 0.0f ,
+                stokKatalis = queryDocumentSnapshot.getLong("stokKatalis")?.toInt() ?: 0 ,
                 komposisi = queryDocumentSnapshot.getString("komposisi")?.split(",") ?: emptyList(),
                 hargaAwal = queryDocumentSnapshot.getLong("hargaAwal")?.toFloat() ?: 0.0f,
                 hargaJual = queryDocumentSnapshot.getLong("hargaJual")?.toFloat() ?: 0.0f,
